fix solaris warnings
Changes to fix warnings on Solaris 5.10.  These changes do not appear to
affect the build on Linux.  Most of these warnings are casting or
including string.h for bzero(), which the Linux man pages agrees with.
I wonder why there is no warning on Linux systems.

+++ src/client.c
@@ -368,7 +368,7 @@ int create_pid_file()
     return ERROR;
   }
 
-  fprintf(fpid,"%d", getpid());
+  fprintf(fpid,"%d", (unsigned int) getpid());
   fclose(fpid);
 
   return OK;
